pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mutli-agent-concierge"
version = "0.2.0"
description = ""
authors = ["Laurie Voss <github@seldo.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
llama-index-core = "^0.11.19"
llama-index-llms-openai = "^0.2.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
llama-index-llms-anthropic = "^0.2.0"
llama-index-agent-openai = "^0.3.0"
llama-index-utils-workflow = "^0.2.2"
guardrails-ai = "^0.6.2"
presidio = "^0.1.0"
presidio-analyzer = "^2.2.357"
